third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/anonymous-iframe/sharedworker-partitioning.tentative.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<body>
<script>
const sw_url = location.pathname.replace(/[^/]*$/, '') +
      "./resources/sharedworker-partitioning-helper.js";

promise_test(async t => {
  // Create 4 iframes (two normal and two anonymous ones) and create
  // a shared worker with the same url in all of them.
  //
  // Creating the same shared worker again with the same url is a
  // no-op. However, anonymous iframes get partitioned shared workers,
  // so we should have a total of 2 shared workers at the end (one for
  // the normal iframes and one for the anonymous ones).
  let iframes = await Promise.all([
    { name: "normal", anonymous: false},
    { name: "normal_control", anonymous: false},
    { name: "anonymous", anonymous: true},
    { name: "anonymous_control", anonymous: true},
  ].map(async ({name, anonymous}) => {

    let iframe = await new Promise(resolve => {
      let iframe = document.createElement('iframe');
      iframe.onload = () => resolve(iframe);
      iframe.src = '/common/blank.html';
      if (anonymous) iframe.anonymous = true;
      document.body.append(iframe);
    });

    let sw = new iframe.contentWindow.SharedWorker(sw_url);
    return { iframe: iframe, name: name, sw: sw };
  }));

  // Ping each worker telling him which iframe it belongs to.
  iframes.forEach((iframe, i) => {
    iframe.sw.port.postMessage({ action: 'record', from: iframe.name });
  });

  let msg_promises = iframes.map(iframe => new Promise(resolve => {
    iframe.sw.port.onmessage = event => resolve(event.data);
  }));

  // Ping each (iframe, sharedworker) asking for which messages it got.
  iframes.map(iframe => iframe.sw.port.postMessage({ action: 'retrieve' }));

  // Collect all replies.
  let msgs = await Promise.all(msg_promises);

  // The "normal" iframe sharedworker belongs to the "normal" and the
  // "normal_control" iframes.
  assert_true(!!msgs[0]["normal"]);
  assert_true(!!msgs[0]["normal_control"]);
  assert_false(!!msgs[0]["anonymous"]);
  assert_false(!!msgs[0]["anonymous_control"]);

  // The "normal_control" iframe shares the same sharedworker as the "normal"
  // iframe.
  assert_true(!!msgs[1]["normal"]);
  assert_true(!!msgs[1]["normal_control"]);
  assert_false(!!msgs[1]["anonymous"]);
  assert_false(!!msgs[1]["anonymous_control"]);

  // The "anonymous" iframe sharedworker belongs to the "anonymous" and the
  // "anonymous_control" iframes.
  assert_false(!!msgs[2]["normal"]);
  assert_false(!!msgs[2]["normal_control"]);
  assert_true(!!msgs[2]["anonymous"]);
  assert_true(!!msgs[2]["anonymous_control"]);

  // The "anonymous_control" iframe shares the same sharedworker as
  // the "anonymous" iframe.
  assert_false(!!msgs[3]["normal"]);
  assert_false(!!msgs[3]["normal_control"]);
  assert_true(!!msgs[3]["anonymous"]);
  assert_true(!!msgs[3]["anonymous_control"]);

}, "Anonymous iframes get partitioned shared workers.");

</script>
</body>
